Week-08-CICD/frontend/requirements.txt
pandas==2.2.3
plotly==5.24.1
gradio==5.7.1
python-dotenv==1.0.1
opencv-python==4.10.0.84